Switch print to logging
Currently, the convention for providers is to invoke print directly. For
code using Lexicon as a library, it would be helpful to use a pattern
that allows for suppressing or redirecting output (e.g., to a log file).

This change replaces all print to DEBUG-level calls to the Python logging
standard module.

Resolves #124