Replace HTTP GET with local file pytest fixture (#408)
* Replace HTTP GET with pytest tmpdir fixture.

Fixes #378.

* Add docstring to testfile fixture.

* Return test runner to original execution directory after completed test.

* Refactor test to use simpler method for the test file.

In this case, the zone file will be created in the directory where tests
are run, and then removed after the tests have completed. One drawback
of this method is that if there is an execution failure before the file
is removed, a developer may not notice, and could accidentally commit
the file to the repository. An entry to .gitignore was added to prevent
this scenario.