Fixed create and update operations for GoDaddy if CAA records are present (#506)
* Fixed create and update operations for GoDaddy if CAA records are present

The GoDaddy API breaks (returning 422) if you try to submit a CAA
record.  So when you create a new record, if the domain has one or more
CAA records, the create will fail (update fails the same way).  This
change simply submits the new (or updated) record instead of submitting
all records.

The logic for locating the record to update remains the same as before,
but now, it won't try to send the update if nothing actually changed,
nor will a message be logged about updating if it didn't attempt to
update.

This doesn't address the problems with delete, as the GoDaddy API still
doesn't let you delete a single record.  So for domains with CAA
records, any delete action will continue to fail.

* Fixed GoDaddy provider to handle things as tests expect

Fixed _create_record to ignore duplicates and handle record sets
appropriately.
Removed some debugging code that was left in create_record by mistake.

Fixed _update_record to handle the potential case of chaning the name of
a record, when update_record updates via identifier, which happens in
the tests.

Fixed _delete_record to work in more cases when CAA records exist, by
updating only the records of a given type.  If the record is identified
by its identifier.  Comment at the top has been updated to reflect the
changes.

* Updated tests for GoDaddy provider to match new functionality.

* Fixed typo in variable name

Also fixed some pylint warnings