Fixed create and update operations for GoDaddy if CAA records are present (#545)
* Fixed create and update operations for GoDaddy if CAA records are present

The GoDaddy API breaks (returning 422) if you try to submit a CAA
record.  So when you create a new record, if the domain has one or more
CAA records, the create will fail (update and delete fail the same way).
This change simply submits the new (or updated) record instead of
submitting all records.

This doesn't address the problems with delete, as the GoDaddy API still
doesn't let you delete a single record.  So for domains with CAA
records, any delete action will continue to fail.

* Fixed typo in  tests/fixtures/cassettes/godaddy/IntegrationTests/test_provider_when_calling_create_record_multiple_times_should_create_record_set.yaml

* Updated test in lines 101 - 105 as requested

This simplifies the conditional checks, since the godaddy API shouldn't
be returning anything that doesn't match the type and name already.

Co-authored-by: Adrien Ferrand <ferrand.ad@gmail.com>